Refactor introduce build tasks
We have now `BuildSpec` and `BuildTask` types. `BuildSpec` describes a build
without a binding to a concrete build config. `BuildTask` is a more concrete
build representation which has env and command rendered for a given config.

That way builders has less API to work with — they get a task a produce side
effects to perform the build.