Resolve & convert opam packages on the fly
Previously we stored a collection of converted opam packages along with
esy.

This commit makes opam resolver fetch opam repo locally into cache
directory and then use to convert packages on-demand.

We still use Python script for conversion which will be replaced later
by a proper code for parsing opam metadata and converting into
package.json format.

For now the result of conversion is suboptimal as we don't consult the
whitelist/blacklist of packages which results in some incompat versions
of packages being resolved. We need a proper opam metadata parser to
filter them out.