Relax req not to have multiple versions of a pkg
Previously if we encountered multiple packages of the same name we
failed. Now we allow that as it is useful for some of the packages which
are needed only at build time and are not linked to any lib/executable.

For example a preprocessor like reason can be used by different packages
and those packages can require different version of reason (even
incompat like v1 and v2).

Of course in case some packages would want to link both versions of the
same package — it will fail at link time.