Add source type to build spec
I add `sourceType: 'transient' | 'immutable'` to `BuildSpec` type. This
is similar to `shouldBePersistent` but not quite.

It provides info if sources for the build are either transient (reside
in some local dir) or immutable (installed from release). Then builders
can decide on their own:

1. if they want to put artefacts in local or global store
2. if they need to rebuild sources on invokation

This commit also adds back the ability for makefile builder to clean &
rebuild sources for `transient` (previously `local`) builds. This is
more correct and now it doesn't hurt so much as we have simple builder
which can decide if it needs to rebuild intelligently (based on
checksums).

The motivation for this change was to have transient sources rebuilt for
releases. Next commit will update the tests cases.